requirements.txt
Django
djangorestframework
celery
requests
beautifulsoup4
redis
selenium
